Move redis queue to mirror app initialiaztion. Fix redissues.
